refactor: move collections into appropriate package
