+++ core/src/main/scala/it/unibo/scafi/xc/engine/context/ContextFactory.scala
@@ -2,5 +2,5 @@ package it.unibo.scafi.xc.engine.context
 
 import it.unibo.scafi.xc.engine.network.Network
 
-trait ContextFactory[N <: Network[_, _, _], C <: Context[_, _, _]]:
+trait ContextFactory[N <: Network[?, ?, ?], C <: Context[?, ?, ?]]:
   def create(network: N): C
