+++ core/src/main/scala/it/unibo/scafi/xc/engine/context/common/OutboundMessagesSemantics.scala
@@ -1,9 +1,10 @@
 package it.unibo.scafi.xc.engine.context.common
 
+import it.unibo.scafi.xc.collections.MapWithDefault
 import scala.collection.{ mutable, MapView }
 
 import it.unibo.scafi.xc.engine.context.Context
-import it.unibo.scafi.xc.engine.network.{ Import, MapWithDefault }
+import it.unibo.scafi.xc.engine.network.Import
 import it.unibo.scafi.xc.engine.path.{ Path, ValueTree }
 
 trait OutboundMessagesSemantics:
@@ -12,11 +13,12 @@ trait OutboundMessagesSemantics:
   type Envelope
 
   override def outboundMessages: Import[DeviceId, InvocationCoordinate, Envelope] =
-    var messages: Map[DeviceId, ValueTree[InvocationCoordinate, Envelope]] = Map.empty
+    var messages: MapWithDefault[DeviceId, ValueTree[InvocationCoordinate, Envelope]] =
+      MapWithDefault.empty(ValueTree.empty)
     for (path, messageMap) <- sentMessages do
       for deviceId <- unalignedDevices do
         messages = messages.updatedWith(deviceId)(_.map(_ + (path -> messageMap(deviceId))))
-    MapWithDefault[DeviceId, ValueTree[InvocationCoordinate, Envelope]](messages, ValueTree.empty)
+    messages
 
   private val sentMessages: mutable.Map[Path[InvocationCoordinate], MapWithDefault[DeviceId, Envelope]] =
     mutable.Map.empty
