+++ core/src/main/scala/it/unibo/scafi/xc/engine/network/Import.scala
@@ -1,5 +1,6 @@
 package it.unibo.scafi.xc.engine.network
 
+import it.unibo.scafi.xc.collections.MapWithDefault
 import it.unibo.scafi.xc.engine.path.ValueTree
 
 type Import[DeviceId, Token, Value] = MapWithDefault[DeviceId, ValueTree[Token, Value]]
